Change from RRTPlanner to use double precision and use old Bezier Planner

+++ soccer/planning/RRTPlanner.hpp
@@ -125,7 +125,7 @@ protected:
         const std::vector<Geometry2d::Point>& points,
         const MotionConstraints& motionConstraints, Geometry2d::Point vi,
         Geometry2d::Point vf,
-        const boost::optional<std::vector<float>>& times = boost::none);
+        const boost::optional<std::vector<double>>& times = boost::none);
 
     /**
      * Generates a velocity profile from a Cubic Bezier Path under the given
